+++ Meshtastic/Helpers/MeshPackets.swift
@@ -1333,7 +1333,7 @@ func waypointPacket (packet: MeshPacket, context: NSManagedObjectContext) {
 				waypoint.longDescription = waypointMessage.description_p
 				waypoint.latitudeI = waypointMessage.latitudeI
 				waypoint.longitudeI = waypointMessage.longitudeI
-				waypoint.icon = Int32(waypointMessage.icon)
+				waypoint.icon = Int64(waypointMessage.icon)
 				waypoint.locked = waypointMessage.locked
 				if waypointMessage.expire != 0 {
 					waypoint.expire = Date(timeIntervalSince1970: TimeInterval(Int64(waypointMessage.expire)))
@@ -1352,7 +1352,7 @@ func waypointPacket (packet: MeshPacket, context: NSManagedObjectContext) {
 				fetchedWaypoint[0].longDescription = waypointMessage.description_p
 				fetchedWaypoint[0].latitudeI = waypointMessage.latitudeI
 				fetchedWaypoint[0].longitudeI = waypointMessage.longitudeI
-				fetchedWaypoint[0].icon = Int32(waypointMessage.icon)
+				fetchedWaypoint[0].icon = Int64(waypointMessage.icon)
 				fetchedWaypoint[0].locked = waypointMessage.locked
 				if waypointMessage.expire != 0 {
 					fetchedWaypoint[0].expire = Date(timeIntervalSince1970: TimeInterval(Int64(waypointMessage.expire)))
